Simplify adding CSS in mobile- and desktop-only scenarios
Previously it was only possible to easily filter for values of
at least a given breakpoint, and even then the names didn't match
up to our common CSS classes for use-cases. This introduces mixins
which match those CSS classes, allowing for filtering in either
direction, and the reimplements those CSS classes in terms of the
new filters to ensure consistency.

As a side-effect this also makes it possible for desktop-only things
to have a `display` of something other than `inline` when present,
as we no longer need to assign a value in such cases (due to only
applying the "hide" case when needed).